(2)4.1 - add file utils.py , fix code
