+++ admin/user_identification_information_admin/urls.py
@@ -5,7 +5,7 @@ from . import views
 app_name = 'admin'
 
 urlpatterns = [
-    url(r'^$', views.UserIdentificationList.as_view(), name='index'),
-    url(r'^(?P<guid>[a-z0-9]+)/$', views.UserIdentificationDetails.as_view(), name='details'),
+    url(r'^$', views.UserIdentificationListAdminView.as_view(), name='index'),
+    url(r'^(?P<guid>[a-z0-9]+)/$', views.UserIdentificationDetailsAdminView.as_view(), name='details'),
 
 ]
